Biggest change: added Google documents highlighter scripts, sans tracking.
This required me to move all existing scripts into a directory scoped to their
respective app.
So the velocity calculation scripts are now located in the spreadsheets/
directory.
The tools/ directory contains a simple 'build.js' script which can be run from
anywhere; it's able to fetch the CDN version of highlight.js and combine it with
a className --> inline styles translation map.